Font update
Today I did a lot of adjustments in the Latin glyphs. In the end of the
day I started creating some kerning classes in Font Forge. But, a
really strange bug happened, excluding a lot of glyphs and reordering
others after that I created some kerning tables. I have my
modifications back in the OTF file, copying the new curves and spaces.
I’ll work the kerning in a separate file to see what happens.